+++ src/components/custom/search/ResultsBlock.jsx
@@ -1,15 +1,14 @@
-import React, {useContext, useEffect, useState} from 'react'
+import React, {useContext, useEffect, useState, useCallback, useRef} from 'react'
 import PropTypes from 'prop-types'
 import {opsDict, ResultsPerPage} from "./ResultsPerPage";
 import DataTable from "react-data-table-component";
-import {
-    PagingInfo
-} from "@elastic/react-search-ui";
-import TableResultsContext from "../../../context/TableResultsContext";
+import TableResultsContext from "@/context/TableResultsContext";
 import ColumnsDropdown from "./ColumnsDropdown";
 import {eq} from '../js/functions'
 import {COLS_ORDER_KEY} from "@/config/config";
 import Spinner from '../Spinner';
+import SearchActions from "@/components/custom/search/SearchActions";
+import {getCheckboxes} from "@/components/custom/BulkExport";
 
 function ResultsBlock({getTableColumns, disableRowClick, tableClassName = '', defaultHiddenColumns = [], searchContext, totalRows, isBusy}) {
 
@@ -31,6 +30,7 @@ function ResultsBlock({getTableColumns, disableRowClick, tableClassName = '', de
         updateTablePagination,
         pageSize,
         pageNumber,
+        raw,
     } = useContext(TableResultsContext)
 
     useEffect(() => {
@@ -39,14 +39,39 @@ function ResultsBlock({getTableColumns, disableRowClick, tableClassName = '', de
         }
     }, [isBusy]);
 
+    const selectedRows = useRef([])
+    const sel = {
+        selectAllIo: 'select-all-rows',
+        selectedCount: 'sui-selected-count'
+    }
 
     const [hiddenColumns, setHiddenColumns] = useState(null)
+
+    const handleRowSelected = useCallback(state => {
+        const $selAllIo =  $(`[name="${sel.selectAllIo}"`)
+
+        const $checkBoxAll = $($selAllIo).parent()
+        $checkBoxAll.find(`.${sel.selectedCount}`).remove()
+
+        if (state.selectedCount) {
+            selectedRows.current = state.selectedRows
+            // add count label
+            $checkBoxAll.append(`<span for="${sel.selectAllIo}" class="${sel.selectedCount}">(${state.selectedCount})</span>`)
+        }
+    }, [])
+
+
+    const rowSelectCriteria = row => {
+        let rows = selectedRows.current.map((e)=> e.id)
+        return rows.indexOf(row.id) !== -1
+    }
 
     return (
         <>
             <div className='sui-layout-main-header'>
                 <div className='sui-layout-main-header__inner'>
-                    <PagingInfo />
+
+                    <SearchActions selectedRows={selectedRows.current} filters={filters} data={getTableData()} raw={raw} hiddenColumns={hiddenColumns} columns={currentColumns.current} />
                     {rows.length > 0 && <ColumnsDropdown searchContext={searchContext} filters={filters} defaultHiddenColumns={defaultHiddenColumns} getTableColumns={getTableColumns} setHiddenColumns={setHiddenColumns}
                                       currentColumns={currentColumns.current} />}
                     <ResultsPerPage updateTablePagination={updateTablePagination}
@@ -57,7 +82,7 @@ function ResultsBlock({getTableColumns, disableRowClick, tableClassName = '', de
             </div>
 
 
-            {<DataTable key={`results-${new Date().getTime()}`}
+            {<DataTable
                         onColumnOrderChange={cols => {
                             currentColumns.current.current = cols
                             const headers = cols.map((col) => eq(typeof col.name, 'string') ? col.name : col.id)
@@ -80,10 +105,15 @@ function ResultsBlock({getTableColumns, disableRowClick, tableClassName = '', de
                         paginationRowsPerPageOptions={Object.keys(opsDict)}
                         pagination
                         paginationServer
+                        //paginationServerOptions={{persistSelectedOnPageChange: true, persistSelectedOnSort: true}}
                         paginationDefaultPage={pageNumber}
                         paginationTotalRows={totalRows || rawResponse.record_count}
                         progressPending={isSearching}
                         progressComponent={<Spinner />}
+                        selectableRowSelected={rowSelectCriteria}
+                        selectableRows
+                        onSelectedRowsChange={handleRowSelected}
+                        selectableRowsVisibleOnly={true}
                 />}
         </>
     )
